requirements.txt
python-dotenv
grpcio
grpcio-tools
protobuf
     